requirements.txt
fastapi
uvicorn
beautifulsoup4
sentence-transformers
faiss-cpu
pdfplumber
python-dotenv
requests

